Add error message, when try to delete category which has some attached data

+++ app/Http/Controllers/Admin/SpotCategoriesController.php
@@ -80,6 +80,10 @@ class SpotCategoriesController extends Controller
      */
     public function destroy($category)
     {
+        if ($category->spots()->withoutNewest()->withRequested()->count() > 0) {
+            return back()->withErrors(['Some spots attached to this category!']);
+        }
+
         $category->delete();
 
         return back();
